+++ tests/phpunit/bootstrap.php
@@ -10,6 +10,16 @@ define( 'DISABLE_SEARCH_PLUGIN_FILE', dirname( __FILE__, 3 ) . '/disable-search.
 ini_set( 'display_errors', 'on' );
 error_reporting( E_ALL );
 
+// Backward compatibility (PHPUnit < 6).
+$phpunit_backcompat = array(
+	'\PHPUnit\Framework\TestCase' => 'PHPUnit_Framework_TestCase',
+);
+foreach ( $phpunit_backcompat as $new => $old ) {
+	if ( ! class_exists( $new ) && class_exists( $old ) ) {
+		class_alias( $old, $new );
+	}
+}
+
 $_tests_dir = getenv( 'WP_TESTS_DIR' );
 if ( ! $_tests_dir ) {
 	$_tests_dir = '/tmp/wordpress-tests-lib';