+++ CHANGELOG.md
@@ -8,6 +8,7 @@
         * Change: Move `phpunit/bin` into `tests/`
     * Change: Remove 'test-' prefix from unit test file
     * Change: In bootstrap, store path to plugin file constant
+    * Change: In bootstrap, add backcompat for PHPUnit pre-v6.0
 
 ## 1.8.3 _(2021-05-01)_
 * Fix: Change `__wakeup()` method visibility from `private` to `public` to avoid warnings under PHP8